+++ dwertheimer.TaskAutomations/__tests__/NPTaskScanAndProcess.test.js
@@ -20,25 +20,25 @@ describe(`${PLUGIN_NAME}`, () => {
   describe(`${FILENAME}`, () => {
     //functions go here using jfunc command
     /*
-     * processUserActionOnLine()
+     * prepareUserAction()
      */
-    describe('processUserActionOnLine()' /* function */, () => {
+    describe('prepareUserAction()' /* function */, () => {
       test('should remove date when userChoice="__remove__"', async () => {
         const origPara = { content: `foo >2020-01-01 bar !!` }
         const changedPara = { content: `foo bar !!` }
-        const result = await f.processUserActionOnLine(origPara, {}, '__remove__')
+        const result = await f.prepareUserAction(origPara, {}, '__remove__')
         expect(result).toEqual({ action: 'set', changed: changedPara })
       })
       test('should remove >1 date when userChoice="__remove__"', async () => {
         const origPara = { content: `foo >2020-01-01 bar !! >2020-01-01` }
         const changedPara = { content: `foo bar !!` }
-        const result = await f.processUserActionOnLine(origPara, {}, '__remove__')
+        const result = await f.prepareUserAction(origPara, {}, '__remove__')
         expect(result).toEqual({ action: 'set', changed: changedPara })
       })
       test('should remove >today when userChoice="__remove__"', async () => {
         const origPara = { content: `foo >today bar !! >2020-01-01` }
         const changedPara = { content: `foo bar !!` }
-        const result = await f.processUserActionOnLine(origPara, {}, '__remove__')
+        const result = await f.prepareUserAction(origPara, {}, '__remove__')
         expect(result).toEqual({ action: 'set', changed: changedPara })
       })
     })