+++ helpers/sorting.js
@@ -1,5 +1,5 @@
 // @flow
-import { get } from 'lodash-es'
+import get from 'lodash/get'
 
 export interface SortableParagraphSubset {
   content: string;